Refactoring to models directory Removal of config from public repo and refresh of API keys
