+++ app/main.py
@@ -1,7 +1,5 @@
-from alpaca_trade_api.entity import Clock
-
-from app.repository.alpaca.alpaca_service import AlpacaService
-from app.repository.alpaca.models.MarketStatus import MarketStatus
+from app.repository.alpaca.alpaca_service import AlpacaService, get_account
+from app.repository.alpaca.models.MarketStatus import print_current_market_status
 
 alpaca = AlpacaService()
 
@@ -11,22 +9,15 @@ def print_nasdaq_assets(active_assets):
     print(nasdaq_assets)
 
 
-def print_market_status(clock):
-    # TODO: Why does this not know what the properties are despite the type hint?
-    mapped = MarketStatus(clock.timestamp, clock.is_open, clock.next_open, clock.next_close)
-    print(mapped)
-
-
-def print_current_market_status():
-    clock_observable = alpaca.get_clock()
-    clock_observable.subscribe(lambda clock: print_market_status(clock))
+def print_account_status():
+    account_observable = get_account()
+    account_observable.subscribe()
 
 
 def main():
     print_current_market_status()
-
-    active_assets_observable = alpaca.get_active_assets()
-
+    print_account_status()
+    #active_assets_observable = alpaca.get_active_assets()
     # active_assets_observable.subscribe(lambda active_assets: print_nasdaq_assets(active_assets))
