+++ app/repository/alpaca/alpaca_service.py
@@ -4,9 +4,12 @@ from typing import Protocol
 from rx import Observable
 from abc import abstractmethod
 from app.config.config import alpaca_key, alpaca_secret
-from app.repository.alpaca.models import MarketStatus
 
 api = tradeapi.REST(alpaca_key, alpaca_secret, api_version='v2')
+
+
+def get_account() -> Observable:
+    return rx.of(api.get_account())
 
 
 class IAlpacaService(Protocol):
@@ -18,6 +21,9 @@ class IAlpacaService(Protocol):
     @abstractmethod
     def get_active_assets(self) -> Observable:
         pass
+
+
+
 
 
 class AlpacaService(IAlpacaService):
